lulua/test_optimize.py
# Copyright (c) 2019 lulua contributors
# 
# Permission is hereby granted, free of charge, to any person obtaining a copy
# of this software and associated documentation files (the "Software"), to deal
# in the Software without restriction, including without limitation the rights
# to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
# copies of the Software, and to permit persons to whom the Software is
# furnished to do so, subject to the following conditions:
# 
# The above copyright notice and this permission notice shall be included in
# all copies or substantial portions of the Software.
# 
# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
# FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
# AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
# LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
# OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
# THE SOFTWARE.

from .optimize import Annealer

class NullAnnealer (Annealer):
    """ Simple dummy annealer for testing """
    def energy (self):
        return sum (self.state)

    def mutate (self):
        prev = self.energy ()
        self.state = [x-1 for x in self.state]
        return self.energy () - prev

def test_null_annealer ():
    dut = NullAnnealer ([1, 2, 3])
    optimal, energy = dut.run (1)
    assert optimal == [0, 1, 2]
    assert energy == sum ([0, 1, 2])-sum([1, 2, 3])
    assert dut.energy () == sum([0, 1, 2])

